+++ src/Event/mt_evtMisc.c
@@ -735,7 +735,7 @@ void FUN_0038c830(int param_1,long param_2);
 void FUN_0038ca00(void);
 void FUN_0038ca80(int param_1);
 void FUN_0038cc10(int param_1,u16 *param_2,int param_3);
-void FUN_0038d560(long param_1,long param_2,u32 param_3,u32 *param_4, u32 *param_5,u32 *param_6);
+void FUN_0038d560(int param_1,int param_2,u32 param_3,u32 *param_4, u32 *param_5,u32 *param_6);
 void FUN_0038d6b0(int param_1,u32 param_2);
 void FUN_0038d720(int param_1);
 void FUN_0038d790(int param_1);
@@ -6830,7 +6830,7 @@ void FUN_0038cc10(int param_1,u16 *param_2,int param_3)
 // FUN_0038D560 NONMATCHING
 
 
-void FUN_0038d560(long param_1,long param_2,u32 param_3,u32 *param_4,
+void FUN_0038d560(int param_1,int param_2,u32 param_3,u32 *param_4,
 
                  u32 *param_5,u32 *param_6)
